Move LSP code actions into helix-view, enable internal actions
This change moves the LSP code actions handling code into helix-view
and introduces a generic action type. This encapsulates all of the LSP
code action details in the new `action` module and paves the way for
future internal actions. (Note that we could alternatively encapsulate
the LSP details in `handlers/lsp.rs`.)

For example, the spell checking integration will be able to write
actions for a spelling mistake like so:

    let mut actions = Vec::new()
    for suggestion in suggestions {
        actions.push(Action::new(
            format!("Replace '{word}' with '{suggestion}'"),
            PRIORITY,
            // This closure's environment would capture a `doc_id`,
            // `view_id`, and the `suggestion`.
            move |editor| {
                // Create and apply a Transaction to replace `word`'s
                // range in the source document with the suggestion
                // string. Clear any diagnostics for that spelling
                // mistake.
                todo!()
            }
        ))
    }
    let word = word.to_string();
    actions.push(Action::new(
        format!("Add '{word}' to the dictionary"),
        PRIORITY,
        // This closure's environment would capture the word. If
        // multiple languages/dictionaries are supported then it would
        // also capture the language.
        move |editor| {
            editor.dictionary.add(&word);
            // Re-check any documents using this dictionary.
            todo!()
        }
    ))

Since all LSP details are now in `actions`, the `code_action` command
has been moved to the general `commands` module - the function is now
only in charge of UI interactions.

This change also makes a patch to the LSP code actions code: all actions
are now combined into a Vec and sorted together. Beforehand, each
language server's code actions were sorted separately and then collected
into one list based on the order in which the language server was
defined.